Build improvements (#2832)
* Build improvements

* fixes 1

* fixes 2

* fixes 3

* fixes 4

* fixes 5

* fixes 6

* fixes 7

* fixes 8

* fixes 1

* fixes 2

* fixes 3

* fixes 4

* fixes 5

* fixes 6

* fixes 7

* fixes 8

* fixes 9

* fixes 10

+++ sbin/pack.sh
@@ -37,12 +37,16 @@ if [[ $1 == --help || $1 == help || $HELP == 1 ]]; then
 		JUST_PRINT=1        Only print package names, do not generate
 		VERBOSE=1           Print commands
 		IGNERR=1            Do not abort on error
+		NOP=1               Print commands, do not execute
 
 	END
 	exit 0
 fi
 
 #----------------------------------------------------------------------------------------------
+
+OP=""
+[[ $NOP == 1 ]] && OP=echo
 
 ARCH=$($READIES/bin/platform --arch)
 [[ $ARCH == x64 ]] && ARCH=x86_64
@@ -83,7 +87,7 @@ fi
 mkdir -p $ARTDIR $ARTDIR/snapshots
 ARTDIR=$(cd $ARTDIR && pwd)
 
-MODULE_NAME=${MODULE_NAME:-redisgraph}
+MODULE_NAME=${MODULE_NAME:-graph}
 PACKAGE_NAME=${PACKAGE_NAME:-redisgraph}
 
 DEP_NAMES=""
@@ -135,45 +139,57 @@ pack_ramp() {
 		$xtx_vars \
 		-e NUMVER -e SEMVER \
 		$RAMP_YAML > /tmp/ramp.yml
-
-	rm -f /tmp/ramp.fname $packfile
-	$RAMP_CMD pack -m /tmp/ramp.yml \
-		$RAMP_ARGS \
-		-n $MODULE_NAME \
-		--verbose \
-		--debug \
-		--packname-file /tmp/ramp.fname \
-		-o $packfile \
-		$MODULE \
-		>/tmp/ramp.err 2>&1 || true
-
-	if [[ ! -e $packfile ]]; then
-		eprint "Error generating RAMP file:"
-		>&2 cat /tmp/ramp.err
-		exit 1
-	else
-		local packname=`cat /tmp/ramp.fname`
-		echo "Created $packname"
+	if [[ $VERBOSE == 1 ]]; then
+		echo "# ramp.yml:"
+		cat /tmp/ramp.yml
 	fi
 
-	if [[ -f $MODULE.debug ]]; then
+	runn rm -f /tmp/ramp.fname $packfile
+	runn @ <<-EOF
 		$RAMP_CMD pack -m /tmp/ramp.yml \
 			$RAMP_ARGS \
 			-n $MODULE_NAME \
 			--verbose \
 			--debug \
 			--packname-file /tmp/ramp.fname \
-			-o $packfile_debug \
-			$MODULE.debug \
+			-o $packfile \
+			$MODULE \
 			>/tmp/ramp.err 2>&1 || true
+		EOF
 
-		if [[ ! -e $packfile_debug ]]; then
+	if [[ $NOP != 1 ]]; then
+		if [[ ! -e $packfile ]]; then
 			eprint "Error generating RAMP file:"
 			>&2 cat /tmp/ramp.err
 			exit 1
 		else
 			local packname=`cat /tmp/ramp.fname`
-			echo "Created $packname"
+			echo "# Created $packname"
+		fi
+	fi
+
+	if [[ -f $MODULE.debug ]]; then
+		runn @ <<-EOF
+			$RAMP_CMD pack -m /tmp/ramp.yml \
+				$RAMP_ARGS \
+				-n $MODULE_NAME \
+				--verbose \
+				--debug \
+				--packname-file /tmp/ramp.fname \
+				-o $packfile_debug \
+				$MODULE.debug \
+				>/tmp/ramp.err 2>&1 || true
+			EOF
+
+		if [[ $NOP != 1 ]]; then
+			if [[ ! -e $packfile_debug ]]; then
+				eprint "Error generating RAMP file:"
+				>&2 cat /tmp/ramp.err
+				exit 1
+			else
+				local packname=`cat /tmp/ramp.fname`
+				echo "# Created $packname"
+			fi
 		fi
 	fi
 
@@ -196,24 +212,36 @@ pack_deps() {
 	local dep_prefix_dir=$(cat $ARTDIR/$dep.prefix)
 	
 	rm -f $tar_path
-	{ cd $depdir ;\
-	  cat $ARTDIR/$dep.files | \
-	  xargs tar -c --sort=name --owner=root:0 --group=root:0 --mtime='UTC 1970-01-01' \
-	  	--transform "s,^,$dep_prefix_dir," 2>> /tmp/pack.err | \
-	  gzip -n - > $tar_path ; E=$?; } || true
-	if [[ ! -e $tar_path || -z $(tar tzf $tar_path) ]]; then
-		eprint "Count not create $tar_path. Aborting."
-		rm -f $tar_path
-		exit 1
+	if [[ $NOP != 1 ]]; then
+		{ cd $depdir ;\
+		  cat $ARTDIR/$dep.files | \
+		  xargs tar -c --sort=name --owner=root:0 --group=root:0 --mtime='UTC 1970-01-01' \
+			--transform "s,^,$dep_prefix_dir," 2>> /tmp/pack.err | \
+		  gzip -n - > $tar_path ; E=$?; } || true
+		if [[ ! -e $tar_path || -z $(tar tzf $tar_path) ]]; then
+			eprint "Count not create $tar_path. Aborting."
+			rm -f $tar_path
+			exit 1
+		fi
+	else
+		runn @ <<-EOF
+			cd $depdir
+			cat $ARTDIR/$dep.files | \
+			xargs tar -c --sort=name --owner=root:0 --group=root:0 --mtime='UTC 1970-01-01' \
+				--transform "s,^,$dep_prefix_dir," 2>> /tmp/pack.err | \
+			gzip -n - > $tar_path ; E=$?; } || true
+			EOF
 	fi
-	sha256sum $tar_path | awk '{print $1}' > $tar_path.sha256
+	runn @ <<-EOF
+		sha256sum $tar_path | awk '{print $1}' > $tar_path.sha256
+		EOF
 
 	mkdir -p $ARTDIR/snapshots
 	cd $ARTDIR/snapshots
 	if [[ ! -z $BRANCH ]]; then
 		local snap_package=$stem.${BRANCH}${VARIANT}.tgz
-		ln -sf ../$fq_package $snap_package
-		ln -sf ../$fq_package.sha256 $snap_package.sha256
+		runn ln -sf ../$fq_package $snap_package
+		runn ln -sf ../$fq_package.sha256 $snap_package.sha256
 	fi
 }
 
@@ -221,12 +249,12 @@ pack_deps() {
 
 prepare_symbols_dep() {
 	if [[ ! -f $MODULE.debug ]]; then return 0; fi
-	echo "Preparing debug symbols dependencies ..."
+	echo "# Preparing debug symbols dependencies ..."
 	echo $(dirname $(realpath $MODULE)) > $ARTDIR/debug.dir
 	echo $(basename $(realpath $MODULE)).debug > $ARTDIR/debug.files
 	echo "" > $ARTDIR/debug.prefix
 	pack_deps debug
-	echo "Done."
+	echo "# Done."
 }
 
 #----------------------------------------------------------------------------------------------
@@ -282,15 +310,15 @@ fi
 #----------------------------------------------------------------------------------------------
 
 if [[ $DEPS == 1 ]]; then
-	echo "Building dependencies ..."
+	echo "# Building dependencies ..."
 
 	[[ $SYM == 1 ]] && prepare_symbols_dep
 
 	for dep in $DEP_NAMES; do
-		echo "$dep ..."
+		echo "# $dep ..."
 		pack_deps $dep
 	done
-	echo "Done."
+	echo "# Done."
 fi
 
 #----------------------------------------------------------------------------------------------
@@ -303,7 +331,7 @@ if [[ $RAMP == 1 ]]; then
 		exit 1
 	fi
 
-	echo "Building RAMP files ..."
+	echo "# Building RAMP files ..."
 
 	[[ -z $MODULE ]] && { eprint "Nothing to pack. Aborting."; exit 1; }
 	[[ ! -f $MODULE ]] && { eprint "$MODULE does not exist. Aborting."; exit 1; }
@@ -312,12 +340,12 @@ if [[ $RAMP == 1 ]]; then
 	[[ $RELEASE == 1 ]] && SNAPSHOT=0 pack_ramp
 	[[ $SNAPSHOT == 1 ]] && pack_ramp
 	
-	echo "Done."
+	echo "# Done."
 fi
 
 if [[ $VERBOSE == 1 ]]; then
-	echo "Artifacts:"
-	du -ah --apparent-size $ARTDIR
+	echo "# Artifacts:"
+	$OP du -ah --apparent-size $ARTDIR
 fi
 
 exit 0
